requirements.txt
Flask==0.12.2
pandas
numpy
Pillow
nltk
opencv-python
tensorflow
requests
MarkupSafe==1.0